requirements.txt
# Django and Core Dependencies
Django==5.2.4
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# File Handling and Media
Pillow==10.1.0
python-magic==0.4.27

# PDF Generation
reportlab==4.0.7
xhtml2pdf==0.2.11

# Excel/CSV Handling
openpyxl==3.1.2
xlsxwriter==3.1.9

# Email and Notifications
django-anymail==10.2
twilio==8.10.0

# Payment Gateway Integration
stripe==7.8.0
requests==2.31.0

# API and Serialization
drf-yasg==1.21.7
django-extensions==3.2.3

# Security and Authentication
django-allauth==0.57.0
django-oauth-toolkit==2.3.0

# Caching and Performance
redis==5.0.1
django-redis==5.4.0
celery==5.3.4

# Task Queue and Background Jobs
celery[redis]==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Monitoring and Logging
sentry-sdk==1.38.0
django-debug-toolbar==4.2.0

# Development and Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2

# Code Quality
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Environment and Configuration
python-decouple==3.8
django-environ==0.11.2

# Date and Time
python-dateutil==2.8.2
pytz==2023.3

# Data Validation
marshmallow==3.20.1
cerberus==1.3.5

# Image Processing
django-imagekit==5.0.0

# Search and Indexing
django-elasticsearch-dsl==8.0

# API Documentation
drf-spectacular==0.26.5

# File Upload and Storage
django-storages==1.14.2
boto3==1.34.0

# Development Server
django-extensions==3.2.3

# Database Migrations
# django-migrate==0.1.0  # Removed - not a valid package

# Form Handling
django-crispy-forms==2.1
crispy-bootstrap5==0.7

# Frontend Integration
django-webpack-loader==0.7.0

# Internationalization
django-modeltranslation==0.18.12

# Admin Interface Enhancement
django-jet==1.0.8
django-grappelli==2.15.1

# REST API Enhancement
django-rest-auth==0.9.5
djangorestframework-simplejwt==5.3.0

# Background Tasks
django-background-tasks==1.2.5

# File Compression
django-compressor==4.4

# Static Files
whitenoise==6.6.0

# Production Server
gunicorn==21.2.0
uvicorn==0.24.0

# Health Checks
django-health-check==3.17.0

# Rate Limiting
django-ratelimit==4.1.0

# Maintenance Mode
django-maintenance-mode==0.15.0

# Error Tracking
# django-sentry==0.1.0  # Removed - not a valid package

# Performance Monitoring
django-silk==5.0.2

# Database Connection Pooling
django-db-connection-pool==1.0.0

# Async Support
django-async==0.1.0

# WebSocket Support
channels==4.0.0
channels-redis==4.1.0

# GraphQL Support (Optional)
graphene-django==3.1.5

# Machine Learning Integration (Optional)
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2

# Data Visualization (Optional)
matplotlib==3.8.2
seaborn==0.13.0

# Testing and Quality Assurance
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.5.0

# Code Coverage
coverage-badge==1.1.0

# Documentation
Sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Deployment
fabric==2.7.1
invoke==2.2.0

# Docker Support
docker-compose==1.29.2

# CI/CD
pre-commit==3.5.0
